alura-school/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
defusedxml==0.7.1
Django==4.1.5
django-admin-honeypot==1.1.0
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
importlib-metadata==6.0.0
Markdown==3.4.1
Pillow==9.4.0
pytz==2022.7
redis==4.4.2
sqlparse==0.4.3
zipp==3.11.0
